==> analytics/0_3_configure_mlflow.ipynb <==
import mlflow

experiment = mlflow.set_experiment(experiment_name='Inzynierka')
print("Experiment_id: {}".format(experiment.experiment_id))
print("Artifact Location: {}".format(experiment.artifact_location))
print("Tags: {}".format(experiment.tags))
print("Lifecycle_stage: {}".format(experiment.lifecycle_stage))
