==> analytics/1_0_get_test_data.ipynb <==
# Załadowanie paczek

import pandas as pd
from sqlalchemy import create_engine
from tqdm import tqdm

## Pobranie danych z bazy

db_con = create_engine('mysql+mysqlconnector://root:password@localhost/CSGOAnalysis?allow_local_infile=1')
db_con

def get_summed_columns(columns):
    both_teams = [
        f"{team}Player_1_{column} + {team}Player_2_{column} + {team}Player_3_{column} + {team}Player_4_{column} + {team}Player_5_{column} as {team}_{column}"
        for column in columns
            for team in ["ct", "t"]
    ]
    team_specific = [
        f"{team}Player_1_{column} + {team}Player_2_{column} + {team}Player_3_{column} + {team}Player_4_{column} + {team}Player_5_{column} as {team}_{column}"
            for team, column in [("ct", "hasDefuse"), ("t", "hasBomb"), ("ct", "isDefusing"), ("t", "isPlanting")]
    ]
    return both_teams + team_specific

get_summed_columns(["hp", "armor", "hasHelmet", "DecoyGrenade", "Flashbang", "HEGrenade", "SmokeGrenade", "fireGrenades", "isBlinded"])

def get_position_columns():
    players = [f"{team}Player_{i}_{pos}"
        for team in ["ct", "t"]
            for i in range(1, 6)
                for pos in ["x", "y","z"]
    ]
    bomb = [f"bomb_{pos}" for pos in ["x", "y","z"]]
    return players + bomb
get_position_columns()

def get_additional_columns(cols):
    additional_cols = [
        f"{team}Player_{i}_{col}"
        for team in ["ct", "t"]
            for i in range(1, 6)
                for col in cols
    ]
    return additional_cols
get_additional_columns(["spotters", "activeWeapon", "mainWeapon", "secondaryWeapon", "isAlive", "lastPlaceName", "ID"])

def build_query():
    selected_columns = get_summed_columns(["hp", "armor", "hasHelmet", "DecoyGrenade", "Flashbang", "HEGrenade", "SmokeGrenade", "fireGrenades", "isBlinded"]) + \
                get_position_columns() + get_additional_columns(["spotters", "activeWeapon", "mainWeapon", "secondaryWeapon", "isAlive", "lastPlaceName", "ID"])
    selected_columns_query = ", ".join(selected_columns)
    query = """SELECT g.filename, g.mapName, r.ctBuyType, r.tBuyType, r.winningSide, f.matchID, f.roundNum,
                    f.seconds, f.ctAlivePlayers, f.ctEqVal, f.tAlivePlayers, f.tEqVal, f.bombsite,
            """ + \
            selected_columns_query + \
            """ FROM frame f
                INNER JOIN round r ON f.matchID=r.matchID AND f.roundNum=r.roundNum
                INNER JOIN game g ON g.ID=f.matchID
                WHERE f.tick < r.endTickCorrect AND (g.filename like "IEM-Katowice%" OR g.filename like "BLAST-Premier-Spring-Final-2022-%")
            """
    return query

query = build_query()
with db_con.connect() as connection:
    states = pd.read_sql(query, con=connection)
states

states.info()

import gc
gc.collect()

states.to_parquet("data/states_test.parquet")
