==> analytics/1_1_prepare_data_v2_test.ipynb <==
# Załadowanie paczek

import pandas as pd
import numpy as np
import json
from awpy.data import PLACE_DIST_MATRIX, NAV
from awpy.analytics.nav import find_closest_area
from sqlalchemy import create_engine
from tqdm import tqdm

# Read parquet

states = pd.read_parquet("data/states_test.parquet")
states

states.info()

## PLACE_DIST

PLACE_DIST_MATRIX.keys()

PLACE_DIST_MATRIX['de_mirage'].keys()

# Modyfikacje

db_con = create_engine('mysql+mysqlconnector://root:password@localhost/CSGOAnalysis?allow_local_infile=1')
db_con

## Odległości do teameatów

def get_dist(mapName, team, spots, areAlive, positions):
    new_positions = []
    for lastPlace, position, isAlive in zip(spots, positions, areAlive):
        if isAlive:
            if lastPlace in PLACE_DIST_MATRIX[mapName]:
                newPlace = lastPlace
            else:
                areaID = find_closest_area(mapName, position)['areaId']
                newPlace = NAV[mapName][areaID]["areaName"]
            new_positions.append(newPlace)
  
    if len(new_positions) <= 1:
        return 9000, 9000
    else:
        dist = [
            PLACE_DIST_MATRIX[mapName][position_x][position_y]['geodesic']['median_dist']
                for i, position_x in enumerate(new_positions)
                    for j, position_y in enumerate(new_positions)
                        if i != j
            ]
        dist_arr = np.array(dist)
        dist_arr = dist_arr[np.isfinite(dist_arr)]
        if dist_arr.size == 0:
            return 9000, 9000
        return np.mean(dist_arr), np.min(dist_arr)

dist_parts = []
for team in tqdm(['ct', 't'], leave=True, position=0):
    data = [get_dist(mapName, team, [spot1, spot2, spot3, spot4, spot5],
                       [isAlive2, isAlive3, isAlive4, isAlive5],
                       [[p1_x, p1_y, p1_z], [p2_x, p2_y, p2_z], [p3_x, p3_y, p3_z], [p4_x, p4_y, p4_z], [p5_x, p5_y, p5_z]])
                for mapName, spot1, spot2, spot3, spot4, spot5, isAlive1, isAlive2, isAlive3, isAlive4, isAlive5,
                    p1_x, p1_y, p1_z, p2_x, p2_y, p2_z, p3_x, p3_y, p3_z, p4_x, p4_y, p4_z, p5_x, p5_y, p5_z in zip(
                                   states[f'mapName'],
                                   states[f'{team}Player_1_lastPlaceName'],
                                   states[f'{team}Player_2_lastPlaceName'],
                                   states[f'{team}Player_3_lastPlaceName'],
                                   states[f'{team}Player_4_lastPlaceName'],
                                   states[f'{team}Player_5_lastPlaceName'],
                                   states[f'{team}Player_1_isAlive'],
                                   states[f'{team}Player_2_isAlive'],
                                   states[f'{team}Player_3_isAlive'],
                                   states[f'{team}Player_4_isAlive'],
                                   states[f'{team}Player_5_isAlive'],
                                   states[f'{team}Player_1_x'],
                                   states[f'{team}Player_1_y'],
                                   states[f'{team}Player_1_z'],
                                   states[f'{team}Player_2_x'],
                                   states[f'{team}Player_2_y'],
                                   states[f'{team}Player_2_z'],
                                   states[f'{team}Player_3_x'],
                                   states[f'{team}Player_3_y'],
                                   states[f'{team}Player_3_z'],
                                   states[f'{team}Player_4_x'],
                                   states[f'{team}Player_4_y'],
                                   states[f'{team}Player_4_z'],
                                   states[f'{team}Player_5_x'],
                                   states[f'{team}Player_5_y'],
                                   states[f'{team}Player_5_z'])] 
    part = pd.DataFrame(data, columns=[f'{team}_meanDist', f'{team}_minDist'])
    dist_parts.append(part)

dist_parts[0]

## Kille

with db_con.connect() as connection:
    kills = pd.read_sql("""
    SELECT k.attackerID, matchID, roundNum, sum(kills) over(partition by k.attackerID, matchID order by roundNum) AS growing_kills
    FROM (
        SELECT attackerID, matchID, roundNum, count(*) as kills
        FROM csgoanalysis.elimination
        WHERE attackerID is NOT NULL
        GROUP BY attackerID, matchID, roundNum
    ) k""", connection)
kills

kills["roundNum"] = kills["roundNum"] + 1
kills

for team in tqdm(['ct', 't']):
    for i in range(1, 6):
        states = states.merge(kills, how='left', left_on=[f'{team}Player_{i}_ID', 'matchID', 'roundNum'], right_on=['attackerID', 'matchID', 'roundNum'],
                             suffixes=(None, f"_{team}_{i}"))
states

states.loc[~states.growing_kills.isna(), 'growing_kills']

states = states.rename(columns={'growing_kills': 'growing_kills_ct_1'})

states.loc[states.roundNum == 1, [f'growing_kills_{team}_{i}' for team in ['ct', 't'] for i in range(1, 6)]] = 0

states = states.drop(columns=states.columns[states.columns.str.startswith('attackerID')])
states

for team in ['ct', 't']: 
    for i in range(1, 6):
        states[f'growing_kills_{team}_{i}'] = states[f'growing_kills_{team}_{i}'].fillna(method='ffill')

## Aktualny wynik

with db_con.connect() as connection:
    rounds = pd.read_sql("""
    SELECT matchID, roundNum, tScore, ctScore
    FROM csgoanalysis.round""", connection)
rounds

states = states.merge(rounds, left_on=['matchID', 'roundNum'], right_on=['matchID', 'roundNum'])
states

## Przejęcie mapy

def get_place(isAlive, mapName, lastPlace, position):
    if isAlive:
        if lastPlace in PLACE_DIST_MATRIX[mapName]:
            newPlace = lastPlace
        else:
            areaID = find_closest_area(mapName, position)['areaId']
            newPlace = NAV[mapName][areaID]["areaName"]
    else:
        newPlace = None
    return newPlace

data = []
for mapName, roundNum, lastRoundNum, \
    ct_spot_1, ct_x_1, ct_y_1, ct_z_1, ct_alive_1, \
    ct_spot_2, ct_x_2, ct_y_2, ct_z_2, ct_alive_2, \
    ct_spot_3, ct_x_3, ct_y_3, ct_z_3, ct_alive_3, \
    ct_spot_4, ct_x_4, ct_y_4, ct_z_4, ct_alive_4, \
    ct_spot_5, ct_x_5, ct_y_5, ct_z_5, ct_alive_5, \
    t_spot_1, t_x_1, t_y_1, t_z_1, t_alive_1, \
    t_spot_2, t_x_2, t_y_2, t_z_2, t_alive_2, \
    t_spot_3, t_x_3, t_y_3, t_z_3, t_alive_3, \
    t_spot_4, t_x_4, t_y_4, t_z_4, t_alive_4, \
    t_spot_5, t_x_5, t_y_5, t_z_5, t_alive_5 in tqdm(zip(states['mapName'], states['roundNum'], states['roundNum'].shift(fill_value=0),
        states['ctPlayer_1_lastPlaceName'], states['ctPlayer_1_x'], states['ctPlayer_1_y'], states['ctPlayer_1_z'], states['ctPlayer_1_isAlive'],
        states['ctPlayer_2_lastPlaceName'], states['ctPlayer_2_x'], states['ctPlayer_2_y'], states['ctPlayer_2_z'], states['ctPlayer_2_isAlive'],
        states['ctPlayer_3_lastPlaceName'], states['ctPlayer_3_x'], states['ctPlayer_3_y'], states['ctPlayer_3_z'], states['ctPlayer_3_isAlive'],
        states['ctPlayer_4_lastPlaceName'], states['ctPlayer_4_x'], states['ctPlayer_4_y'], states['ctPlayer_4_z'], states['ctPlayer_4_isAlive'],
        states['ctPlayer_5_lastPlaceName'], states['ctPlayer_5_x'], states['ctPlayer_5_y'], states['ctPlayer_5_z'], states['ctPlayer_5_isAlive'],
        states['tPlayer_1_lastPlaceName'], states['tPlayer_1_x'], states['tPlayer_1_y'], states['tPlayer_1_z'], states['tPlayer_1_isAlive'],
        states['tPlayer_2_lastPlaceName'], states['tPlayer_2_x'], states['tPlayer_2_y'], states['tPlayer_2_z'], states['tPlayer_2_isAlive'],
        states['tPlayer_3_lastPlaceName'], states['tPlayer_3_x'], states['tPlayer_3_y'], states['tPlayer_3_z'], states['tPlayer_3_isAlive'],
        states['tPlayer_4_lastPlaceName'], states['tPlayer_4_x'], states['tPlayer_4_y'], states['tPlayer_4_z'], states['tPlayer_4_isAlive'],
        states['tPlayer_5_lastPlaceName'], states['tPlayer_5_x'], states['tPlayer_5_y'], states['tPlayer_5_z'], states['tPlayer_5_isAlive']), 
                                                     total=len(states.index)):
    if lastRoundNum != roundNum:
        ct_areas = set()
        t_areas = set()
        
    curr_ct_areas = {
        get_place(ct_alive_1, mapName, ct_spot_1, [ct_x_1, ct_y_1, ct_z_1]),
        get_place(ct_alive_2, mapName, ct_spot_2, [ct_x_2, ct_y_2, ct_z_2]),
        get_place(ct_alive_3, mapName, ct_spot_3, [ct_x_3, ct_y_3, ct_z_3]),
        get_place(ct_alive_4, mapName, ct_spot_4, [ct_x_4, ct_y_4, ct_z_4]),
        get_place(ct_alive_5, mapName, ct_spot_5, [ct_x_5, ct_y_5, ct_z_5])
    }
    curr_t_areas = {
        get_place(t_alive_1, mapName, t_spot_1, [t_x_1, t_y_1, t_z_1]),
        get_place(t_alive_2, mapName, t_spot_2, [t_x_2, t_y_2, t_z_2]),
        get_place(t_alive_3, mapName, t_spot_3, [t_x_3, t_y_3, t_z_3]),
        get_place(t_alive_4, mapName, t_spot_4, [t_x_4, t_y_4, t_z_4]),
        get_place(t_alive_5, mapName, t_spot_5, [t_x_5, t_y_5, t_z_5])
    }

    for area in curr_ct_areas:
        if area not in curr_t_areas:
            ct_areas.add(area)
        if area in t_areas:
            t_areas.remove(area)

    for area in curr_t_areas:
        if area not in curr_ct_areas:
            t_areas.add(area)
        if area in ct_areas:
            ct_areas.remove(area)
    
    ct_areas.discard(None)
    t_areas.discard(None)
    ct_percentage = len(ct_areas)/len(PLACE_DIST_MATRIX[mapName].keys())
    t_percentage = len(t_areas)/len(PLACE_DIST_MATRIX[mapName].keys()) 
    data.append((ct_percentage, t_percentage))
    
conquer = pd.DataFrame.from_records(data, columns=['ct_conquerMap', 't_conquerMap'])

conquer

## Zgrupowanie graczy

def get_groupped_players(mapName, spots, alive, positions):
    counter = {}
    for spot, isAlive, position in zip(spots, alive, positions):
        if isAlive:
            if spot in PLACE_DIST_MATRIX[mapName]:
                newPlace = spot
            else:
                areaID = find_closest_area(mapName, position)['areaId']
                newPlace = NAV[mapName][areaID]["areaName"]
            counter[newPlace] = counter.get(newPlace, 0) + 1
    maximum = 0 if len(counter.values()) == 0 else max(counter.values())
    return maximum

groupped_parts = []
for team in tqdm(['ct', 't'], leave=True, position=0):
    data = [ 
        get_groupped_players(mapName=mapName,
                            spots=[spot1, spot2, spot3, spot4, spot5], 
                            alive=[isAlive1, isAlive2, isAlive3, isAlive4, isAlive5],
                            positions=[[p1_x, p1_y, p1_z], [p2_x, p2_y, p2_z], [p3_x, p3_y, p3_z], [p4_x, p4_y, p4_z], [p5_x, p5_y, p5_z]])
                for mapName, spot1, spot2, spot3, spot4, spot5, isAlive1, isAlive2, isAlive3, isAlive4, isAlive5,
                    p1_x, p1_y, p1_z, p2_x, p2_y, p2_z, p3_x, p3_y, p3_z, p4_x, p4_y, p4_z, p5_x, p5_y, p5_z in 
                    zip(
                       states[f'mapName'],
                       states[f'{team}Player_1_lastPlaceName'],
                       states[f'{team}Player_2_lastPlaceName'],
                       states[f'{team}Player_3_lastPlaceName'],
                       states[f'{team}Player_4_lastPlaceName'],
                       states[f'{team}Player_5_lastPlaceName'],
                       states[f'{team}Player_1_isAlive'],
                       states[f'{team}Player_2_isAlive'],
                       states[f'{team}Player_3_isAlive'],
                       states[f'{team}Player_4_isAlive'],
                       states[f'{team}Player_5_isAlive'],
                       states[f'{team}Player_1_x'],
                       states[f'{team}Player_1_y'],
                       states[f'{team}Player_1_z'],
                       states[f'{team}Player_2_x'],
                       states[f'{team}Player_2_y'],
                       states[f'{team}Player_2_z'],
                       states[f'{team}Player_3_x'],
                       states[f'{team}Player_3_y'],
                       states[f'{team}Player_3_z'],
                       states[f'{team}Player_4_x'],
                       states[f'{team}Player_4_y'],
                       states[f'{team}Player_4_z'],
                       states[f'{team}Player_5_x'],
                       states[f'{team}Player_5_y'],
                       states[f'{team}Player_5_z'])
        ]
    part = pd.DataFrame(data, columns=[f'{team}_grouppedPlayers'])
    groupped_parts.append(part)

## Odległości do bombsite'u

def get_dist(isAlive, mapName, lastPlace, position):
    if isAlive:
        if lastPlace in PLACE_DIST_MATRIX[mapName]:
            newPlace = lastPlace
        else:
            areaID = find_closest_area(mapName, position)['areaId']
            newPlace = NAV[mapName][areaID]["areaName"]
        distToA = PLACE_DIST_MATRIX[mapName][newPlace]['BombsiteA']['geodesic']['median_dist']
        distToB = PLACE_DIST_MATRIX[mapName][newPlace]['BombsiteB']['geodesic']['median_dist']
    else:
        distToA = None
        distToB = None
    return distToA, distToB

all_parts = []
for team in tqdm(['ct', 't'], leave=True, position=0):
    for i in tqdm(range(1, 6), leave=True, position=1):
        data = [ 
            get_dist(isAlive, mapName, lastPlace, [x, y, z])
                for isAlive, mapName, lastPlace, x, y, z in 
                    zip(states[f'{team}Player_{i}_isAlive'],
                        states[f'mapName'], 
                        states[f'{team}Player_{i}_lastPlaceName'],
                        states[f'{team}Player_{i}_x'],
                        states[f'{team}Player_{i}_y'],
                        states[f'{team}Player_{i}_z'])
        ]
        part = pd.DataFrame.from_records(data, columns=[f'{team}Player_{i}_distToA', f'{team}Player_{i}_distToB'])
        all_parts.append(part)

## Widziani gracze

def get_spotted_players(*args):
    spotted = set()
    for arg in args:
        spotted.update(json.loads(arg))
    return len(spotted)

spotters_parts = []
for team in tqdm(['ct', 't'], leave=True, position=0):
    data = [ 
        get_spotted_players(spot1, spot2, spot3, spot4, spot5)
                for spot1, spot2, spot3, spot4, spot5 in 
                    zip(states[f'{team}Player_1_spotters'],
                       states[f'{team}Player_2_spotters'],
                       states[f'{team}Player_3_spotters'],
                       states[f'{team}Player_4_spotters'],
                       states[f'{team}Player_5_spotters'])
        ]
    part = pd.DataFrame(data, columns=[f'{team}_spottedPlayers'])
    spotters_parts.append(part)

# Łączenie i generowanie

states.drop(columns=states.columns[states.columns.str.endswith(("_x", "_y", "_z", "_ID", "_lastPlaceName", "_spotters"))], inplace=True)
states = pd.concat([states, *dist_parts, *spotters_parts, *all_parts, *groupped_parts, conquer], axis=1)
states.columns.values

def prepare_features(unprep_df):
    pistols = {'Glock-18', 'USP-S', 'P2000', 'P250', 'Dual Berettas'}
    enhanced_pistols = {'CZ75 Auto', 'Five-SeveN', 'Tec-9', 'R8 Revolver'}
    deagle = 'Desert Eagle'
    shotguns = {'MAG-7', 'XM1014', 'Nova', 'Sawed-Off'}
    machine_guns = {'M249', 'Negev'}
    smgs = {'MP9', 'MP7', 'MP5-SD', 'MAC-10', 'UMP-45', 'PP-Bizon', 'P90'}
    weaker_rifles = {'Galil AR', 'SSG 08', 'FAMAS'}
    lunet_rifles = {'SG 553', 'AUG'}
    sniper_rifle = {'G3SG1', 'SCAR-20', 'AWP'}
    assault_rifle = {'M4A1', 'M4A4', 'AK-47'}
    others = {'Zeus x27', 'Knife', 'C4', 'Molotov', 'Incendiary Grenade', 
              'Smoke Grenade', 'Flashbang', 'Decoy Grenade', 'HE Grenade'}
    others.update(shotguns, machine_guns)
    
    def convert_weapons(col):
        for team in ['ct', 't']:
            unprep_df[f"{team}_{col}_Pistol"] = 0
            unprep_df[f"{team}_{col}_EnhancedPistols"] = 0
            unprep_df[f"{team}_{col}_Deagle"] = 0
            unprep_df[f"{team}_{col}_SMG"] = 0
            unprep_df[f"{team}_{col}_WeakAssaultRifle"] = 0
            unprep_df[f"{team}_{col}_LunetRifle"] = 0
            unprep_df[f"{team}_{col}_SniperRifle"] = 0
            unprep_df[f"{team}_{col}_AssaultRifle"] = 0
            for i in range(1, 6):
                unprep_df[f"{team}_{col}_Pistol"] += unprep_df[f"{team}Player_{i}_{col}"].isin(pistols).astype(int)
                unprep_df[f"{team}_{col}_EnhancedPistols"] = unprep_df[f"{team}Player_{i}_{col}"].isin(enhanced_pistols).astype(int)
                unprep_df[f"{team}_{col}_Deagle"] += (unprep_df[f"{team}Player_{i}_{col}"] == deagle).astype(int)
                unprep_df[f"{team}_{col}_SMG"] += unprep_df[f"{team}Player_{i}_{col}"].isin(smgs).astype(int)
                unprep_df[f"{team}_{col}_WeakAssaultRifle"] += unprep_df[f"{team}Player_{i}_{col}"].isin(weaker_rifles).astype(int)
                unprep_df[f"{team}_{col}_LunetRifle"] += unprep_df[f"{team}Player_{i}_{col}"].isin(lunet_rifles).astype(int)
                unprep_df[f"{team}_{col}_SniperRifle"] += unprep_df[f"{team}Player_{i}_{col}"].isin(sniper_rifle).astype(int)
                unprep_df[f"{team}_{col}_AssaultRifle"] += unprep_df[f"{team}Player_{i}_{col}"].isin(assault_rifle).astype(int)
                                
    def convert_active_weapons(col):
        for team in ['ct', 't']:
            unprep_df[f"{team}_{col}_Pistol"] = 0
            unprep_df[f"{team}_{col}_EnhancedPistols"] = 0
            unprep_df[f"{team}_{col}_Deagle"] = 0
            unprep_df[f"{team}_{col}_SMG"] = 0
            unprep_df[f"{team}_{col}_WeakAssaultRifle"] = 0
            unprep_df[f"{team}_{col}_LunetRifle"] = 0
            unprep_df[f"{team}_{col}_SniperRifle"] = 0
            unprep_df[f"{team}_{col}_AssaultRifle"] = 0
            unprep_df[f"{team}_{col}_Others"] = 0
            for i in range(1, 6):
                unprep_df[f"{team}_{col}_Pistol"] += unprep_df[f"{team}Player_{i}_{col}"].isin(pistols).astype(int)
                unprep_df[f"{team}_{col}_EnhancedPistols"] = unprep_df[f"{team}Player_{i}_{col}"].isin(enhanced_pistols).astype(int)
                unprep_df[f"{team}_{col}_Deagle"] += (unprep_df[f"{team}Player_{i}_{col}"] == deagle).astype(int)
                unprep_df[f"{team}_{col}_SMG"] += unprep_df[f"{team}Player_{i}_{col}"].isin(smgs).astype(int)
                unprep_df[f"{team}_{col}_WeakAssaultRifle"] += unprep_df[f"{team}Player_{i}_{col}"].isin(weaker_rifles).astype(int)
                unprep_df[f"{team}_{col}_LunetRifle"] += unprep_df[f"{team}Player_{i}_{col}"].isin(lunet_rifles).astype(int)
                unprep_df[f"{team}_{col}_SniperRifle"] += unprep_df[f"{team}Player_{i}_{col}"].isin(sniper_rifle).astype(int)
                unprep_df[f"{team}_{col}_AssaultRifle"] += unprep_df[f"{team}Player_{i}_{col}"].isin(assault_rifle).astype(int)
                unprep_df[f"{team}_{col}_Others"] += unprep_df[f"{team}Player_{i}_{col}"].isin(others).astype(int)
                
    def convert_sec_weapons(col):
        for team in ['ct', 't']:
            unprep_df[f"{team}_{col}_Pistol"] = 0
            unprep_df[f"{team}_{col}_EnhancedPistols"] = 0
            unprep_df[f"{team}_{col}_Deagle"] = 0
            for i in range(1, 6):
                unprep_df[f"{team}_{col}_Pistol"] += unprep_df[f"{team}Player_{i}_{col}"].isin(pistols).astype(int)
                unprep_df[f"{team}_{col}_EnhancedPistols"] = unprep_df[f"{team}Player_{i}_{col}"].isin(enhanced_pistols).astype(int)
                unprep_df[f"{team}_{col}_Deagle"] += (unprep_df[f"{team}Player_{i}_{col}"] == deagle).astype(int)


    convert_active_weapons('activeWeapon')
    convert_weapons('mainWeapon')
    convert_sec_weapons('secondaryWeapon')
    
    unprep_df['winningSideCT'] = np.where(unprep_df['winningSide'] == 'CT', 1, 0)
    unprep_df['bombsite'] = np.where(unprep_df['bombsite'] == '', 'not_planted', unprep_df['bombsite'])
    unprep_df['ctMinDistToA'] = unprep_df[["ctPlayer_1_distToA", "ctPlayer_2_distToA", "ctPlayer_3_distToA", "ctPlayer_4_distToA", "ctPlayer_5_distToA"]].min(axis=1)
    unprep_df['ctMinDistToB'] = unprep_df[["ctPlayer_1_distToB", "ctPlayer_2_distToB", "ctPlayer_3_distToB", "ctPlayer_4_distToB", "ctPlayer_5_distToB"]].min(axis=1)
    unprep_df['tMinDistToA'] = unprep_df[["tPlayer_1_distToA", "tPlayer_2_distToA", "tPlayer_3_distToA", "tPlayer_4_distToA", "tPlayer_5_distToA"]].min(axis=1)
    unprep_df['tMinDistToB'] = unprep_df[["tPlayer_1_distToB", "tPlayer_2_distToB", "tPlayer_3_distToB", "tPlayer_4_distToB", "tPlayer_5_distToB"]].min(axis=1)
    unprep_df['ctMeanDistToA'] = unprep_df[["ctPlayer_1_distToA", "ctPlayer_2_distToA", "ctPlayer_3_distToA", "ctPlayer_4_distToA", "ctPlayer_5_distToA"]].mean(axis=1)
    unprep_df['ctMeanDistToB'] = unprep_df[["ctPlayer_1_distToB", "ctPlayer_2_distToB", "ctPlayer_3_distToB", "ctPlayer_4_distToB", "ctPlayer_5_distToB"]].mean(axis=1)
    unprep_df['tMeanDistToA'] = unprep_df[["tPlayer_1_distToA", "tPlayer_2_distToA", "tPlayer_3_distToA", "tPlayer_4_distToA", "tPlayer_5_distToA"]].mean(axis=1)
    unprep_df['tMeanDistToB'] = unprep_df[["tPlayer_1_distToB", "tPlayer_2_distToB", "tPlayer_3_distToB", "tPlayer_4_distToB", "tPlayer_5_distToB"]].mean(axis=1)
    unprep_df['ctMeanKills'] = (unprep_df['growing_kills_ct_1'] * unprep_df['ctPlayer_1_isAlive'] + unprep_df['growing_kills_ct_2'] * unprep_df['ctPlayer_2_isAlive'] + 
            unprep_df['growing_kills_ct_3'] * unprep_df['ctPlayer_3_isAlive'] + unprep_df['growing_kills_ct_4'] * unprep_df['ctPlayer_4_isAlive'] + 
            unprep_df['growing_kills_ct_5'] * unprep_df['ctPlayer_5_isAlive']) / unprep_df[[f'ctPlayer_{i}_isAlive' for i in range(1, 6)]].sum(axis=1)
    unprep_df['tMeanKills'] = (unprep_df['growing_kills_t_1'] * unprep_df['tPlayer_1_isAlive'] + unprep_df['growing_kills_t_2'] * unprep_df['tPlayer_2_isAlive'] + 
            unprep_df['growing_kills_t_3'] * unprep_df['tPlayer_3_isAlive'] + unprep_df['growing_kills_t_4'] * unprep_df['tPlayer_4_isAlive'] + 
            unprep_df['growing_kills_t_5'] * unprep_df['tPlayer_5_isAlive']) / unprep_df[[f'tPlayer_{i}_isAlive' for i in range(1, 6)]].sum(axis=1)
    unprep_df['tMeanKills'] = unprep_df['tMeanKills'].fillna(-1)
    unprep_df = pd.concat([unprep_df, pd.get_dummies(unprep_df['mapName']), pd.get_dummies(unprep_df['bombsite'], prefix='bombsite')], axis=1)
    unprep_df['tMinDistToA'] = unprep_df['tMinDistToA'].fillna(9000)
    unprep_df['tMinDistToB'] = unprep_df['tMinDistToB'].fillna(9000)
    unprep_df['tMeanDistToA'] = unprep_df['tMeanDistToA'].fillna(9000)
    unprep_df['tMeanDistToB'] = unprep_df['tMeanDistToB'].fillna(9000)
    return unprep_df

states = prepare_features(states)
states

states['ctMeanKills']

states.loc[states['tMeanKills'].isna(), 'tMeanKills']

states.drop(columns=states.columns[states.columns.str.endswith(("_activeWeapon", "_mainWeapon", "_secondaryWeapon", "_distToA", "_distToB",
                                                               "_isAlive")) | states.columns.str.startswith("growing_kills")],
            inplace=True)
states.columns

states.to_parquet("data/states_another_mod_test.parquet")
